Celý řetězec render() od WorldView až po ObjectInstance jede na floatech a při vykreslování mají zapnutou bilineární interpolaci (aby se obrázky mohly vykreslovat i na necelé pixely). Také je k tomu potřeba AffineTransform neboť drawImage nemá x/y ve floatech. Umí to irotace, strech a další. Rotaci jsem nakonec taky zahrnul a to ve formě houpání – kombinace pootočení = sin(čas)
Celou "platformu" jsem zkusil osvobodit od závislosti na UPDATES_FREQ => vzniky statické metody na ZFrame, které přepošítávají Hz na updateCount a naopak.
Možno vyzkoušet: při změně UPDATES_FREQ mezi 3 a 300 by nemělo být moc poznat. Nízká hodnota ale omezuje, kdy může pták změnit frame (updateCount je celočíselný) => při frekvenci 0.1 by mohl mrknout jen jednou za 10 vteřin apod.

Na mapě small.txt je hodně ptáků. Ptáci se inicializují trochu jinak - i rychlost a flatterSpeed jsou random.

+snažim se uploadovat jen s uklizeným WhiteSpace...

+++ src/net/trdlo/zelda/GameObjectInstance.java
@@ -9,34 +9,34 @@ import java.awt.Graphics2D;
 
 
 public abstract class GameObjectInstance implements Comparable<GameObjectInstance> {
-    protected float posX, posY;
+	protected float posX, posY;
 
-    public GameObjectInstance(float posX, float posY) {
-        this.posX = posX;
-        this.posY = posY;
-    }
+	public GameObjectInstance(float posX, float posY) {
+		this.posX = posX;
+		this.posY = posY;
+	}
 	
 	public void update() {
 		
 	}
 
-    public float getPosX() {
-        return posX;
-    }
+	public float getPosX() {
+		return posX;
+	}
 
-    public float getPosY() {
-        return posY;
-    }
+	public float getPosY() {
+		return posY;
+	}
 
-    public float getMoveX() {
-        return 0;
-    }
+	public float getMoveX() {
+		return 0;
+	}
 
-    public float getMoveY() {
-        return 0;
-    }
-	
-    public abstract void render(Graphics2D graphics, int x, int y, float renderFraction);
+	public float getMoveY() {
+		return 0;
+	}
+
+	public abstract void render(Graphics2D graphics, float x, float y, float renderFraction);
 	
 	public abstract int getZIndex();
